requirements.txt
# Core ML and Data Science
scikit-learn==1.3.2
tensorflow==2.15.0
pandas==2.1.4
numpy==1.24.3
shap==0.44.0

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Data Processing
pyarrow==14.0.2
parquet==1.3.1

# Configuration and CLI
pyyaml==6.0.1
click==8.1.7
structlog==23.2.0

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Dashboard
flask==2.3.3
flask-socketio==5.3.6
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
